ltc: fix no-std, clippy, docs, and audit CI gates
Pull litecoin no_std Vec fix via git tip, import alloc::Vec in MWEB PSBT
helpers, clear rustdoc/clippy failures, exclude nested vendor workspace
from clippy, and ignore known transitive rustls-webpki advisories.

+++ crates/mweb/tests/lip0006_p2p.rs
@@ -39,7 +39,7 @@ fn tcp_p2p_header_and_pmmr_syncs_owned_balance() {
         .expect("Core sendtoaddress mweb");
     env.mine_mweb_activation(&mining).expect("activate");
 
-    let tip_height = env.rpc.get_block_count().unwrap() as u32;
+    let tip_height = env.rpc.get_block_count().unwrap();
     let tip_hash = env.rpc.get_block_hash(tip_height).unwrap();
 
     let mut peer = TcpMwebPeer::connect(env.p2p_addr(), Network::Regtest).expect("p2p connect");
